FileSystem.py: Python 3 compatibility fixes
We can list and read files with Python 2 and 3, using latin and
non-latin characters in path names.

Internally we keep all paths as regular str strings. This means that
in Python 2 the path names are stored as UTF-8 encoded byte arrays,
and in Python 3 the path names are stored as unicode strings.

python-fuse library 1.0.0 and up is *required* for Python 3, see:
https://github.com/libfuse/python-fuse/issues/13